Refactoring Auth System And Models

+++ src/app/Http/Controllers/Admin/PhotoChangeController.php
@@ -3,43 +3,47 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Models\Admin\Admin;
+use App\Models\User;
 use Illuminate\Http\Request;
-use Illuminate\Validation\Rule;
-use DB;
+use Ramsey\Uuid\Uuid;
 
 class PhotoChangeController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('admin');
+        $this->middleware('auth:web');
     }
 
     public function index()
     {
-        $admin_data = Admin::where('id',session('id'))->first();
+        $admin_data = User::where('id', auth()->user()->id)->first();
         return view('admin.auth.photo_change', compact('admin_data'));
     }
 
     public function update(Request $request)
     {
+        $fileName = '';
+
         $request->validate([
             'photo' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048'
         ]);
 
         // Unlink old photo
-        unlink(public_path('uploads/'.$request->current_photo));
+        if($request->current_photo) {
+            unlink(public_path('uploads/'.$request->current_photo));
+            preg_match('/(user-)(.*).(jpg|png|jpeg|gif)/', $request->current_photo, $user_photo_split);
+            $fileName = $user_photo_split[1] . $user_photo_split[2] . '.' . $request->file('photo')->getClientOriginalExtension();
+        } else {
+            $fileName = 'user-' . Uuid::uuid4() . '.' . $request->file('photo')->getClientOriginalExtension();
+        }
 
         // Uploading new photo
-        $ext = $request->file('photo')->extension();
-        $final_name = 'user-'.session('id').'.'.$ext;
-        $request->file('photo')->move(public_path('uploads/'), $final_name);
+        
+        $request->file('photo')->move(public_path('uploads/'), $fileName);
 
-        $data['photo'] = $final_name;
+        $data['photo'] = $fileName;
 
-        Admin::where('id',session('id'))->update($data);
-
-        session(['photo' => $final_name]);
+        User::where('id',auth()->user()->id)->update($data);
 
         return redirect()->back()->with('success', 'Photo is updated successfully!');
